Support using host binutils. Try that for GCC 4.7

+++ gcc/build/build.sh
@@ -79,23 +79,27 @@ applyPatchesAndConfig gcc${VERSION}
 
 echo "Will configure with ${CONFIG}"
 
-echo "Fetching binutils ${BINUTILS_VERSION}"
-if [[ ! -e binutils-${BINUTILS_VERSION}.tar.bz2 ]]; then
-    curl -L -O http://ftp.gnu.org/gnu/binutils/binutils-${BINUTILS_VERSION}.tar.bz2
-fi
-BINUTILS_DIR=binutils-${BINUTILS_VERSION}
-rm -rf ${BINUTILS_DIR}
-tar jxf binutils-${BINUTILS_VERSION}.tar.bz2
-BINUTILS_FILES=$(cd ${BINUTILS_DIR}; ls -1)
-pushd gcc-${VERSION}
-for file in ${BINUTILS_FILES}
-do
-    if [ ! -e "$file" ]
-    then
-        ln -sf "../${BINUTILS_DIR}/${file}"
+if [[ -z "${BINUTILS_VERSION}" ]]; then
+    echo "Using host binutils $(ld -v)"
+else
+    echo "Fetching binutils ${BINUTILS_VERSION}"
+    if [[ ! -e binutils-${BINUTILS_VERSION}.tar.bz2 ]]; then
+        curl -L -O http://ftp.gnu.org/gnu/binutils/binutils-${BINUTILS_VERSION}.tar.bz2
     fi
-done
-popd
+    BINUTILS_DIR=binutils-${BINUTILS_VERSION}
+    rm -rf ${BINUTILS_DIR}
+    tar jxf binutils-${BINUTILS_VERSION}.tar.bz2
+    BINUTILS_FILES=$(cd ${BINUTILS_DIR}; ls -1)
+    pushd gcc-${VERSION}
+    for file in ${BINUTILS_FILES}
+    do
+        if [ ! -e "$file" ]
+        then
+            ln -sf "../${BINUTILS_DIR}/${file}"
+        fi
+    done
+    popd
+fi
 
 echo "Downloading prerequisites"
 pushd gcc-${VERSION}
@@ -109,13 +113,15 @@ make -j$(nproc)
 make install-strip
 popd
 
-# Work around insane in-tree built ld issue
-for bindir in ${STAGING_DIR}/{,x86_64-linux-gnu/}bin
-do
-    pushd ${bindir}
-    ln -s ld ld-new
-    popd
-done
+if [[ -z "${BINUTILS_VERSION}" ]]; then
+    # Work around insane in-tree built ld issue
+    for bindir in ${STAGING_DIR}/{,x86_64-linux-gnu/}bin
+    do
+        pushd ${bindir}
+        ln -s ld ld-new
+        popd
+    done
+fi
 
 # Compress all the images with upx
 upx --best ${STAGING_DIR}/bin/* || true
